post_build: Fixed kernel version detection
PROD-10750

+++ board/post_build_common_60.sh
@@ -3,7 +3,7 @@
 # Copyright (C) 2024 Ezurio
 
 # enable tracing and exit on errors
-set -x -e
+set -x -e -o pipefail
 
 BOARD_DIR="${1}"
 BUILD_TYPE="${2}"
@@ -187,8 +187,10 @@ fi
 [ -d "${TARGET_DIR}/usr/lib/node_modules" ] && \
 	find "${TARGET_DIR}/usr/lib/node_modules" -name '*.md' -exec rm -f {} \;
 
-rm -rf "${TARGET_DIR}/usr/share/gobject-introspection-1.0/"
-rm -rf "${TARGET_DIR}/usr/lib/gobject-introspection/"
+if ! grep -qF "BR2_PACKAGE_GOBJECT_INTROSPECTION=y" "${BR2_CONFIG}"; then
+	rm -rf "${TARGET_DIR}/usr/share/gobject-introspection-1.0/"
+	rm -rf "${TARGET_DIR}/usr/lib/gobject-introspection/"
+fi
 
 rm -rf "${TARGET_DIR}/var/www/swupdate"
 rm -f "${TARGET_DIR}/usr/lib/swupdate/conf.d/90-start-progress"
@@ -292,9 +294,8 @@ esac
 
 "${BR2_EXTERNAL_SUMMIT_SOM_PATH}/board/kernel-fitimage.sh" "${BINARIES_DIR}/kernel.its"
 
-
-printf '%s\n%s\n' "${kver}" '6.6.0' | sort --sort=version | tail -n1 | \
-	grep -qF '6.6.0' && OLD_KERNEL=true || OLD_KERNEL=false
+printf '%s\n%s\n' "${kver}" '6.6.0' | sort --sort=version | head -n1 | \
+	grep -qF '6.6.0' && OLD_KERNEL=false || OLD_KERNEL=true
 
 "${BR2_EXTERNAL_SUMMIT_SOM_PATH}/board/generate_boot_script.sh" \
 	"${BUILD_TYPE}" "${OLD_KERNEL}" "${SECURE_BOOT}" "${ENCRYPTED_TOOLKIT}" \
